Create `bundle_as_identifier` template property
This allows the Android template to rely on briefcase to convert
`-` in bundles into `_`, allowing them to be used as Android app
identifiers and Java package identifiers.

This addresses a problem where apps with `-` in their `bundle`
name, e.g., TravelTips, would not launch properly with the
briefcase Android tooling.