Improve visual appearance of dmg
At the moment, the app icon and application folder symlink are rather large. This PR reduces the icon size to 64 px and adapts the icon positions and window size accordingly. In the medium term, it may be nice to provide options to customise the installer appearance in the pyproject.toml.