Prevent invalid app names on windows, show warning on others
This code (line 123) prevents an app being called invalid filenames such as PRN or CON on windows, which would result in a long error not easily understandable. Also adds a warning for mac users who call their file these names, as this makes downloading to windows impossible - the download is stuck in the middle and gives no error message with some downloading methods like Drive.