Added option to run ADB in no-streaming mode
ADB can install an app either directly from the host ("streamed install"), or
by copying the APK to the device and then running the installation process. The
former is the current default, but the latter is still needed for some devices.
Added an option allowing streamed install to be disabled.

Signed-off-by: Helio Perroni Filho <xperroni@gmail.com>