Appease branch coverage detection
Swapping the "else" case to the most general (generic, default hint) makes this safe. If httpx introduces a fourth RequestError, it will not be obvious in the code, but at least the hint will be generic enough to not confuse the user, and vague enough to perhaps prompt them to check the logs and find out what the underlying error was.