Fix inconsistent naming and perform validity check
Specifying multiple values in a LSItemContentTypes key is only valid
when multiple document types are manually grouped together in the
Info.plist file. For Briefcase apps, document types are always
separately declared in the configuration file, so only a single value
must be provided. Because the key name is plural and is an array in the
Info.plist file, a list *is* accepted, but must have only one element.