Raise ValueError if PEP639 license-files is not list and make tests pass
Without this change, the code will work even if the license-files field is just a single string.
However, in that case, the string will be iterated over and the code will look for a glob match for each character independently.
This fix makes it explicitly fail in that circumstance, as the PEP639 standard specifies that license-files is an array of strings.

Co-authored-by: Yngve Mardal Moe <yngve.m.moe@gmail.com>